+++ docs/changelog.md
@@ -14,6 +14,7 @@ category: Administration
 * test: remove unused mock hints from test case and simplify test
 * docs: add developer documentation regarding unique simple name for actions (#115)
 * refactor: add unique prefix to active sessions action (#115)
+* refactor: add unique prefix to global configuration action (#115)
 
 ### [20.06.0] - 2020-06-13